test: extend virtqueue split tests for event-idx and irq suppression

+++ drivers/windows/virtio/common/virtio_osdep.h
@@ -30,7 +30,7 @@ typedef uint32_t UINT32;
 typedef uint64_t UINT64;
 
 typedef int32_t NTSTATUS;
-typedef int BOOLEAN;
+typedef UINT8 BOOLEAN;
 typedef void VOID;
 
 #ifndef TRUE
